nanobot/requirements-api.txt
# Butler API dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.0.0

# Database (shared with nanobot tools)
asyncpg>=0.29.0

# HTTP client (used by HomeAssistantTool)
aiohttp>=3.9.0

# Claude API
anthropic>=0.40.0

# Auth
PyJWT>=2.8.0
livekit-api>=0.7.0
